+++ include/VirtualMachine.hpp
@@ -39,8 +39,8 @@ namespace StackVM
         int32_t GetDestinationRegisterIndex(Instruction* instruction) const;        
         void SetDestinationRegisterDataType(Instruction* instruction, Type type);
         void IncrementInstructionPointer();
-        void IncrementInstructionPointer(uint32_t offset);
-        void SetInstructionPointer(uint32_t offset);
+        void IncrementInstructionPointer(uint64_t offset);
+        void SetInstructionPointer(uint64_t offset);
         void ResetState();
         void LogMessage(const std::string& message);
     public:
